Add footer, other minor edits
