+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ers-stack",
+  "name": "Knowted",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
